% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accdb_parse}
\alias{accdb_parse}
\title{Convert from database accdb to parsed accnr-tibble}
\usage{
accdb_parse(accdb_str)
}
\arguments{
\item{accdb_str}{The AccNR str from the database}
}
\value{
A parsed accnr_tib
}
\description{
Convert from database accdb to parsed accnr-tibble
}
