% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accdb_parse_to_accnr}
\alias{accdb_parse_to_accnr}
\title{Convert from database accnr to human-readable accnr}
\usage{
accdb_parse_to_accnr(accdb_str)
}
\arguments{
\item{accdb_str}{The AccNR str from the database}
}
\value{
A parsed accnr_list
}
\description{
Convert from database accnr to human-readable accnr
}
