% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accdb_to_accnr}
\alias{accdb_to_accnr}
\title{Convert from database accdb to accnr}
\usage{
accdb_to_accnr(accdb_str)
}
\arguments{
\item{accdb_str}{The AccNR str from the database}
}
\value{
An accnr str
}
\description{
Convert from database accdb to accnr
}
