% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accdb_validate}
\alias{accdb_validate}
\title{Validate AccDB entered by the user}
\usage{
accdb_validate(accdb_str)
}
\arguments{
\item{accdb_str}{The AccDB to be validated}
}
\value{
Logical value wether the accdb_str is a valid accdb that can be parsed or not.
}
\description{
Validate a accdb, on the form 202200000
}
