% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accnr_add}
\alias{accnr_add}
\title{Add a value onto an AccNR}
\usage{
accnr_add(accnr_list, amount)
}
\arguments{
\item{accnr_list}{The AccNR to be validated}

\item{amount}{The value which will be added. Normally a 1 to get the next AccNR}
}
\description{
Given a parsed AccNR, add to the value of it
}
