% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accnr_sprint}
\alias{accnr_sprint}
\title{Format an accnr letter and value to the standard format}
\usage{
accnr_sprint(accnr_list)
}
\arguments{
\item{accnr_list}{The AccNR to format}
}
\value{
A string of the nicely formated AccNR
}
\description{
Format the parsed accnr_list for printing.
}
\examples{
accnr_str <- "A2022/12354"
accnr_list <- accnr_parse(accnr_str)
accnr_next <- accnr_add(accnr_list, 3)
print(accnr_sprint(accnr_next))
}
