% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accnr_to_accdb}
\alias{accnr_to_accdb}
\title{Convert from accnr to database accdb}
\usage{
accnr_to_accdb(accnr_str)
}
\arguments{
\item{accnr_str}{The AccNR str}
}
\value{
An accdb str
}
\description{
Convert from accnr to database accdb
}
