% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accnr_to_database_format}
\alias{accnr_to_database_format}
\title{Convert from human-readable accnr to the accnr saved in the db}
\usage{
accnr_to_database_format(accnr_tib)
}
\arguments{
\item{accnr_tib}{The AccNR to format}
}
\value{
A string of the accnr formated for the db
}
\description{
Convert from human-readable accnr to the accnr saved in the db
}
