% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{disconnect_from_database}
\alias{disconnect_from_database}
\title{Disconnect from database}
\usage{
disconnect_from_database(conn)
}
\arguments{
\item{conn}{A database connection returned by \link[esbaser]{connect_to_database}}
}
\description{
Close connection to database
}
