% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender}
\alias{gender}
\title{Exported table from ESBase}
\description{
Exported table (with same name) from ESBase and loaded into a tibble for easier development without an SQL-server.
}
\references{
\url{http://esbase.nrm.se}
}
\keyword{data}
