% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_accnr_biologdata}
\alias{get_accnr_biologdata}
\title{Get Biologdata from AccNR}
\usage{
get_accnr_biologdata(accnr_str)
}
\arguments{
\item{accnr_str}{The AccNR to get the data for}
}
\value{
All biologdata that already exists for the AccNR
}
\description{
Given an AccNR, get the biologdata that already exists in the DB
If the accnr_str is empty it will return an row of NA/empty values
}
