% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_age}
\alias{get_age}
\title{Get Age}
\usage{
get_age(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire age table from the database
}
