% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{get_analysis_type}
\alias{get_analysis_type}
\title{Get Analysis Type}
\usage{
get_analysis_type(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire analysis_type table from the database
}
