% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_biologdata_colnames}
\alias{get_biologdata_colnames}
\title{Get columnnames of biologdata}
\usage{
get_biologdata_colnames()
}
\value{
The colnames of the biologdata
}
\description{
Get columnnames of biologdata
}
