% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_catalog}
\alias{get_catalog}
\title{Get Catalog}
\usage{
get_catalog(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire catalog table from the database
}
