% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_coast}
\alias{get_coast}
\title{Get Coasts}
\usage{
get_coast(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire coast table from the database
}
