% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_country}
\alias{get_country}
\title{Get Countrys}
\usage{
get_country(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire country table from the database
}
