% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{get_county}
\alias{get_county}
\title{Get Countys}
\usage{
get_county(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire county table from the database
}
