% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_gender}
\alias{get_gender}
\title{Get Gender}
\usage{
get_gender(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire gender table from the database
}
