% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_material_storage}
\alias{get_material_storage}
\title{Get Material Storange}
\usage{
get_material_storage(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire material_storage table from the database
}
