% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{get_material_type}
\alias{get_material_type}
\title{Get Material Type}
\usage{
get_material_type(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire material_type table from the database
}
