% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{get_nourishment}
\alias{get_nourishment}
\title{Get Nourishment}
\usage{
get_nourishment(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire nourishment table from the database
}
