% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_options_lokaler}
\alias{get_options_lokaler}
\title{Get options lokaler}
\usage{
get_options_lokaler()
}
\value{
Tibble with id, representation
}
\description{
Returns the possble options for the locality
}
