% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_options_material_type}
\alias{get_options_material_type}
\title{Get options material_type}
\usage{
get_options_material_type()
}
\value{
Tibble with id, representation
}
\description{
Returns the possble options for the material_type
}
