% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_old.R
\name{get_options_project}
\alias{get_options_project}
\title{Get options project}
\usage{
get_options_project()
}
\value{
Tibble with id, representation
}
\description{
Returns the possble options for the project
}
