% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_old.R
\name{get_options_species}
\alias{get_options_species}
\title{Get options species}
\usage{
get_options_species()
}
\value{
Tibble with id, representation
}
\description{
Returns the possble options for the species
}
