% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_person}
\alias{get_person}
\title{Get Person}
\usage{
get_person(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire person table from the database
}
