% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{get_project}
\alias{get_project}
\title{Get Project}
\usage{
get_project(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire project table from the database
}
