% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{get_province}
\alias{get_province}
\title{Get Provinces}
\usage{
get_province(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire province table from the database
}
