% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{get_species}
\alias{get_species}
\title{Get Species}
\usage{
get_species(conn)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}
}
\value{
A tibble
}
\description{
Get the entire species table from the database
}
