% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_stodlista_arter}
\alias{get_stodlista_arter}
\title{Get stödlista med arter}
\usage{
get_stodlista_arter()
}
\value{
Dataframe med kolumner: Katalog, svenskt namn, latinskt namn, engelskt namn
}
\description{
Returns a helplist containing all species that can be used for checking data input
}
