% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_stodlista_lokaler}
\alias{get_stodlista_lokaler}
\title{Get stödlista med lokaler}
\usage{
get_stodlista_lokaler()
}
\value{
Dataframe with omrade, narmsta_ort, sjodistrikt, landskap, lan, land
}
\description{
Returns a helplist containing all loakler that can be used for checking data input
}
