% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{insert_analysisrecord}
\alias{insert_analysisrecord}
\title{Insert analysisrecord}
\usage{
insert_analysisrecord(
  conn,
  account_id,
  project_id,
  creator_id,
  contact_id,
  date,
  shippingdate,
  analysis_type_id,
  result,
  analysis_type_note
)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}

\item{account_id}{The id of the logged in user to be put in the created_by,updated_by column. Must be a numeric.}

\item{project_id}{The id of the project this analysis is connected to}

\item{creator_id}{The person_id of the creator of the analysis}

\item{contact_id}{The person_id of the contact for this analysis}

\item{date}{Provberedningsdatum}

\item{shippingdate}{The date the sample was sent}

\item{analysis_type_id}{Based on analysis_type}

\item{result}{Notis, comment about the results of the analysis}

\item{analysis_type_note}{Analystypnotis, comment about the analysis_type}
}
\value{
The number of affected rows as affected_rows and the id of the newly created analysisrecord as new_row_id
}
\description{
Insert new analysisrecord
}
