% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{insert_analysisrecord_row}
\alias{insert_analysisrecord_row}
\title{Insert analysisrecord_row}
\usage{
insert_analysisrecord_row(conn, amount, homogenate_amount, tag)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}

\item{amount}{The size of the sample}

\item{homogenate_amount}{If a homogenate, the partial weight}

\item{tag}{The provid of the sample}
}
\value{
The number of affected rows as affected_rows and the id of the newly created analysisrecord as new_row_id
}
\description{
Insert new analysisrecord_row
}
