% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{insert_analysisrecord_row_assn}
\alias{insert_analysisrecord_row_assn}
\title{Insert analysisrecord_row_assn}
\usage{
insert_analysisrecord_row_assn(conn, analysisrecord_id, analysisrecord_row_id)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}

\item{analysisrecord_id}{The id of the newly inserted analysisrecord}

\item{analysisrecord_row_id}{The id of the newly inserted analysisrecord_row}
}
\value{
The number of affected rows as affected_rows in a list
}
\description{
Insert new analysisrecord_row_assn, connect two ids of analysisrecord and analysisrecord_row
}
