% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{insert_new_material}
\alias{insert_new_material}
\title{Insert new material}
\usage{
insert_new_material(conn, account_id, accdbs, material_type_id)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}

\item{account_id}{The id of the logged in user to be put in the created_by,updated_by column. Must be a numeric.}

\item{accdbs}{The accession ids in database form to add to point the materials to}

\item{material_type_id}{The inital material_type_id for the new material rows}
}
\value{
The number of affected rows as affected_rows in a list
}
\description{
Insert new material rows
}
