% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{load_accessions}
\alias{load_accessions}
\title{Get Accession rows from AccNRs}
\usage{
load_accessions(accnr_start, accnr_end)
}
\arguments{
\item{accnr_start}{The accnr to start load from}

\item{accnr_end}{The accnr to end at}
}
\value{
All accession rows between the specified accnrs
}
\description{
Given a vector of AccNRs, get the accession rows
}
