% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load All the Data}
\usage{
load_data(path)
}
\arguments{
\item{path}{The path to the csv file}
}
\value{
Parsed data
}
\description{
Load all the data and do some magic with it.
}
\examples{
data <- load_data("~/data.csv")
}
