% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{parse_accnr}
\alias{parse_accnr}
\title{Parse AccNR in string form}
\usage{
parse_accnr(accnr_str)
}
\arguments{
\item{accnr_str}{The AccNR to be parsed, either with or without an '/'}
}
\value{
A list containing the 'letter', 'year' and numeric 'value' of the accnr
}
\description{
If it is valid a accnr, generate a list containing 'letter', 'year' and 'value' based on it.
}
