% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_old.R
\name{paste_collapse}
\alias{paste_collapse}
\title{Paste Collapse}
\usage{
paste_collapse(x, collapse = ", ")
}
\arguments{
\item{x}{The vector which non-Na non-empty-string values should be pasted and collapsed}

\item{collapse}{Passed to the paste function}
}
\value{
A string of the collapsed values with the collapse param between ignoring empty strings and NAs
}
\description{
Removes all NA:s and empty strings from x, then pastes and collapses them together
}
