% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provid_utils.R
\name{provid_add}
\alias{provid_add}
\title{Add a value onto an ProvID}
\usage{
provid_add(provid_tib, amount)
}
\arguments{
\item{provid_tib}{The ProvID to add onto}

\item{amount}{The value which will be added. Normally a 1 to get the next ProvID}
}
\description{
Given a parsed ProvID, add to the value of it
}
