% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provid_utils.R
\name{provid_validate}
\alias{provid_validate}
\title{Validate ProvID entered by the user}
\usage{
provid_validate(provid_str)
}
\arguments{
\item{provid_str}{The ProvID to be validated}
}
\value{
Logical value wether the provid_str is a valid provid that can be parsed or not.
}
\description{
Validate a provid, on the form Q2022-12345
}
