% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{sprint_accnr}
\alias{sprint_accnr}
\title{Format an accnr letter and value to the standard format}
\usage{
sprint_accnr(accnr_list)
}
\arguments{
\item{accnr_list}{The AccNR to be parsed}
}
\value{
A string of the nicely formated AccNR
}
\description{
Format the parsed accnr_list for printing.
}
\examples{
accnr_str <- "A2022/12354"
accnr_list <- parse_accnr(accnr_str)
accnr_next <- add_accnr(accnr_list, 3)
print(sprint_accnr(accnr_next))
}
