% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{update_fish}
\alias{update_fish}
\title{Update fish rows}
\usage{
update_fish(conn, account_id, fish)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}

\item{account_id}{The id of the logged in user to be put in the created_by,updated_by column}

\item{fish}{A tibble of the fish, containing the columns 'accession_id, nourishment_id, gender_id, liverweight, totallength,
decay_id, reproduction_phase_id, othernumber, bodylength}
}
\value{
The number of affected rows
}
\description{
Update fish rows based on moified tibble returned by \link[esbaser]{get_fish_between} based on the accession_id
}
