% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{update_material}
\alias{update_material}
\title{Update material rows}
\usage{
update_material(conn, account_id, material)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}

\item{account_id}{The id of the logged in user to be put in the created_by,updated_by column}

\item{material}{A tibble of the material, containing the columns 'id, material_type_id, amount_original,
amount_left, storage_id, storage_type_id, storage_note.}
}
\value{
The number of affected rows as affected_rows in a list
}
\description{
Update material rows based on moified tibble returned by \link[esbaser]{get_material_between} based on the id
}
