% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{update_specimen}
\alias{update_specimen}
\title{Update specimen rows}
\usage{
update_specimen(conn, account_id, specimen)
}
\arguments{
\item{conn}{The database connection returned by \link[esbaser]{connect_to_database}}

\item{account_id}{The id of the logged in user to be put in the created_by,updated_by column}

\item{specimen}{A tibble of the specimen, containing the columns 'id, note, storagenote, deathdate_start, deathdate_end,
age_type_id, age_start, age_end, weight}
}
\value{
The number of affected rows
}
\description{
Update material rows based on moified tibble returned by \link[esbaser]{get_specimen_between} based on the accession_id
}
