% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accnr_parse}
\alias{accnr_parse}
\title{Parse AccNR in string form}
\usage{
accnr_parse(accnr_str)
}
\arguments{
\item{accnr_str}{The AccNR to be parsed (single or vector), either with or without an '/'}
}
\value{
A tibble with the columns 'letter', 'year' and numeric 'value' of the accnrs
}
\description{
If it is valid a accnr, generate a list containing 'letter', 'year' and 'value' based on it.
}
