% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accnr_utils.R
\name{accnr_sprint}
\alias{accnr_sprint}
\title{Format an accnr letter and value to the standard format}
\usage{
accnr_sprint(accnr_tib)
}
\arguments{
\item{accnr_tib}{The AccNR to format}
}
\value{
A string of the nicely formated AccNR
}
\description{
Format the parsed accnr_list for printing.
}
\examples{
accnr_str <- c("A2022/12354", "B2023/00001")
accnr_tib <- accnr_parse(accnr_str)
accnr_next <- accnr_add(accnr_tib, c(1, 3))
print(accnr_sprint(accnr_next))
}
