% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Connect to database}
\usage{
connect_to_database(dbname = "test")
}
\arguments{
\item{dbname}{The name of the database, is 'test' on my local computer, 'mgg2' on the production machine}
}
\value{
conn The database connection
}
\description{
Initializes connection to database.
}
