% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_biologdata_colnames}
\alias{get_biologdata_colnames}
\title{Get columnnames of biologdata}
\usage{
get_biologdata_colnames(pretty)
}
\arguments{
\item{pretty}{TRUE or FALSE, wether to return pretty values, 'Lever kvar (g)' or 'lever_kvar'}
}
\value{
The colnames of the biologdata
}
\description{
Get columnnames of biologdata
}
