% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{join_part}
\alias{join_part}
\title{Join Part}
\usage{
join_part(
  tib,
  repr_column,
  id_column = "id",
  cols = c("code", "swe_name"),
  collapse = "-"
)
}
\arguments{
\item{tib}{A tibble contaning an id column, and one or more columns that should be joined into a string representation}

\item{repr_column}{The string of the column-name the representation column should take}

\item{id_column}{The name of the id column in the tibble tib}

\item{cols}{A vector with the columns in tib that should be joined}

\item{collapse}{The string that should separate the columns cols}
}
\value{
A tibble taking the id_column from tib, and the combined cols separated by collapse
}
\description{
Used to compile helplists where for example the species table has a catalog_id, which ponts to the catalog table
}
